added feed components

+++ discordClone/base/views.py
@@ -93,7 +93,10 @@ def home(request):
         Q(description__icontains=q))
     topics = Topic.objects.all()
     room_count = rooms.count()
-    context={'rooms':rooms,'topics':topics,'room_count':room_count}
+    room_messages = Message.objects.all().filter(Q(room__topic__name__icontains =q))
+
+
+    context={'rooms':rooms,'topics':topics,'room_count':room_count,'room_messages':room_messages}
     return render(request,'base/home.html',context)
 
 def room(request,pk):
@@ -107,12 +110,23 @@ def room(request,pk):
             room = room,
             body = request.POST.get('body')
         )
+        room.participants.add(request.user)
+
         return redirect('room',pk =room.id)
 
     context = {'room':room,'room_messages':room_messages,'participants':participants}
 
     return render(request,'base/room.html',context)
 @login_required(login_url = 'login')
+
+
+def userProfile(request,pk):
+    user = User.objects.get(id=pk)
+    rooms = user.room_set.all()
+    context = {'user':user,'rooms':rooms}
+    return render(request, 'base/profile.html',context)
+
+
 def createRoom(request):
     form=RoomForm()
     if request.method=='POST':
@@ -145,7 +159,21 @@ def updateRoom(request,pk):
 @login_required(login_url = 'login')
 def deleteRoom(request,pk):
     room=Room.objects.get(id=pk)
+
+    if request.user !=room.host:
+        return HttpResponse('Your are not allowed here!!')
     if request.method== 'POST':
         room.delete()
         return redirect('home')
     return render(request,'base/delete.html',{'obj':room})
+
+
+@login_required(login_url = 'login')
+def deleteMessage(request,pk):
+    message=Message.objects.get(id=pk)
+    if request.user !=message.user:
+        return HttpResponse('Your are not allowed here!!')
+    if request.method== 'POST':
+        message.delete()
+        return redirect('home')
+    return render(request,'base/delete.html',{'obj':message})
